pybackend2/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

numpy
opencv-python-headless

pymongo

insightface
onnxruntime
